refactor: teachers
