+++ candle/api/api.py
@@ -5,7 +5,8 @@ Author: Daniel Grohol, FMFI UK
 
 
 from flask import Blueprint, request, jsonify
-from candle.models import Teacher, Lesson, teacher_lessons
+from candle.models import Lesson
+from candle.teachers.models import teacher_lessons, Teacher
 from candle.subjects.models import Subject
 from candle.groups.models import StudentGroup
 from candle.rooms.models import Room
