+++ candle/panel/templates/panel/panel.html
@@ -16,7 +16,7 @@
         <div class="panel_cast_obsah">
             {% include('search/teacher_search.html') %}
             <div class="other_links">
-                <a href="{{ url_for('teacher.list_teachers') }}">Rozvrhy všetkých učiteľov</a>
+                <a href="{{ url_for('teachers.listing') }}">Rozvrhy všetkých učiteľov</a>
             </div>
         </div>
     </div>
