+++ candle/timetable/templates/timetable/list.html
@@ -21,7 +21,7 @@
         <td>
             {%- for teacher in lesson.teachers -%}
                 {%- if teacher.slug -%}
-                    <a href="{{ url_for('teacher.show_timetable', slug=teacher.slug) }}">{{ teacher.short_name }}</a>
+                    <a href="{{ url_for('teachers.show_timetable', slug=teacher.slug) }}">{{ teacher.short_name }}</a>
                     {%- if not loop.last -%}
                         {{ ', ' }}
                     {%- endif -%}
